Move ErgoAmount, TokenAmount, PaymentRequest to other package, improve unit tests
