+++ app/src/main/java/org/ergoplatform/android/transactions/ReceiveToWalletFragment.kt
@@ -9,7 +9,9 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.navArgs
 import kotlinx.coroutines.launch
-import org.ergoplatform.android.*
+import org.ergoplatform.android.AppDatabase
+import org.ergoplatform.android.NodeConnector
+import org.ergoplatform.android.R
 import org.ergoplatform.android.databinding.FragmentReceiveToWalletBinding
 import org.ergoplatform.android.ui.copyStringToClipboard
 import org.ergoplatform.android.ui.formatFiatToString
@@ -21,6 +23,7 @@ import org.ergoplatform.android.wallet.addresses.ChooseAddressListDialogFragment
 import org.ergoplatform.android.wallet.addresses.getAddressLabel
 import org.ergoplatform.android.wallet.getDerivedAddress
 import org.ergoplatform.android.wallet.getDerivedAddressEntity
+import org.ergoplatform.getExplorerPaymentRequestAddress
 
 
 /**
