+++ app/src/main/res/layout/fragment_prompt_signing_dialog.xml
@@ -9,7 +9,7 @@
     <ImageView
         android:id="@+id/qr_code"
         android:layout_width="match_parent"
-        android:layout_height="250dp"
+        android:layout_height="325dp"
         android:layout_marginVertical="@dimen/half_horizontal_margin"
         android:contentDescription="@string/label_qr_code"
         tools:src="@drawable/ic_qr_code_24" />
