+++ common-jvm/src/main/java/org/ergoplatform/ErgoAmount.kt
@@ -1,4 +1,4 @@
-package org.ergoplatform.android
+package org.ergoplatform
 
 import java.math.RoundingMode
