+++ common-jvm/src/main/java/org/ergoplatform/TokenAmount.kt
@@ -1,4 +1,4 @@
-package org.ergoplatform.android
+package org.ergoplatform
 
 import kotlin.math.pow
